feat(cohere-+-clarifai): refactor integrations to use common base config class
