+++ litellm/llms/clarifai/common_utils.py
@@ -1,0 +1,8 @@
+import httpx
+
+from litellm.llms.base_llm.transformation import BaseLLMException
+
+
+class ClarifaiError(BaseLLMException):
+    def __init__(self, status_code: int, message: str):
+        super().__init__(status_code=status_code, message=message)
