+++ litellm/llms/cohere/chat/handler.py
@@ -10,9 +10,12 @@ import httpx  # type: ignore
 import requests  # type: ignore
 
 import litellm
+from litellm.litellm_core_utils.litellm_logging import Logging as LiteLLMLoggingObj
 from litellm.litellm_core_utils.streaming_handler import CustomStreamWrapper
+from litellm.llms.base_llm.transformation import BaseConfig, BaseLLMException
 from litellm.llms.custom_httpx.http_handler import AsyncHTTPHandler, HTTPHandler
 from litellm.types.llms.cohere import ToolResultObject
+from litellm.types.llms.openai import AllMessageValues
 from litellm.types.utils import (
     ChatCompletionToolCallChunk,
     ChatCompletionUsageBlock,
@@ -20,120 +23,8 @@ from litellm.types.utils import (
 )
 from litellm.utils import Choices, Message, ModelResponse, Usage
 
-from ..prompt_templates.factory import cohere_message_pt, cohere_messages_pt_v2
-
-
-class CohereError(Exception):
-    def __init__(self, status_code, message):
-        self.status_code = status_code
-        self.message = message
-        self.request = httpx.Request(method="POST", url="https://api.cohere.ai/v1/chat")
-        self.response = httpx.Response(status_code=status_code, request=self.request)
-        super().__init__(
-            self.message
-        )  # Call the base class constructor with the parameters it needs
-
-
-class CohereChatConfig:
-    """
-    Configuration class for Cohere's API interface.
-
-    Args:
-        preamble (str, optional): When specified, the default Cohere preamble will be replaced with the provided one.
-        chat_history (List[Dict[str, str]], optional): A list of previous messages between the user and the model.
-        generation_id (str, optional): Unique identifier for the generated reply.
-        response_id (str, optional): Unique identifier for the response.
-        conversation_id (str, optional): An alternative to chat_history, creates or resumes a persisted conversation.
-        prompt_truncation (str, optional): Dictates how the prompt will be constructed. Options: 'AUTO', 'AUTO_PRESERVE_ORDER', 'OFF'.
-        connectors (List[Dict[str, str]], optional): List of connectors (e.g., web-search) to enrich the model's reply.
-        search_queries_only (bool, optional): When true, the response will only contain a list of generated search queries.
-        documents (List[Dict[str, str]], optional): A list of relevant documents that the model can cite.
-        temperature (float, optional): A non-negative float that tunes the degree of randomness in generation.
-        max_tokens (int, optional): The maximum number of tokens the model will generate as part of the response.
-        k (int, optional): Ensures only the top k most likely tokens are considered for generation at each step.
-        p (float, optional): Ensures that only the most likely tokens, with total probability mass of p, are considered for generation.
-        frequency_penalty (float, optional): Used to reduce repetitiveness of generated tokens.
-        presence_penalty (float, optional): Used to reduce repetitiveness of generated tokens.
-        tools (List[Dict[str, str]], optional): A list of available tools (functions) that the model may suggest invoking.
-        tool_results (List[Dict[str, Any]], optional): A list of results from invoking tools.
-        seed (int, optional): A seed to assist reproducibility of the model's response.
-    """
-
-    preamble: Optional[str] = None
-    chat_history: Optional[list] = None
-    generation_id: Optional[str] = None
-    response_id: Optional[str] = None
-    conversation_id: Optional[str] = None
-    prompt_truncation: Optional[str] = None
-    connectors: Optional[list] = None
-    search_queries_only: Optional[bool] = None
-    documents: Optional[list] = None
-    temperature: Optional[int] = None
-    max_tokens: Optional[int] = None
-    k: Optional[int] = None
-    p: Optional[int] = None
-    frequency_penalty: Optional[int] = None
-    presence_penalty: Optional[int] = None
-    tools: Optional[list] = None
-    tool_results: Optional[list] = None
-    seed: Optional[int] = None
-
-    def __init__(
-        self,
-        preamble: Optional[str] = None,
-        chat_history: Optional[list] = None,
-        generation_id: Optional[str] = None,
-        response_id: Optional[str] = None,
-        conversation_id: Optional[str] = None,
-        prompt_truncation: Optional[str] = None,
-        connectors: Optional[list] = None,
-        search_queries_only: Optional[bool] = None,
-        documents: Optional[list] = None,
-        temperature: Optional[int] = None,
-        max_tokens: Optional[int] = None,
-        k: Optional[int] = None,
-        p: Optional[int] = None,
-        frequency_penalty: Optional[int] = None,
-        presence_penalty: Optional[int] = None,
-        tools: Optional[list] = None,
-        tool_results: Optional[list] = None,
-        seed: Optional[int] = None,
-    ) -> None:
-        locals_ = locals()
-        for key, value in locals_.items():
-            if key != "self" and value is not None:
-                setattr(self.__class__, key, value)
-
-    @classmethod
-    def get_config(cls):
-        return {
-            k: v
-            for k, v in cls.__dict__.items()
-            if not k.startswith("__")
-            and not isinstance(
-                v,
-                (
-                    types.FunctionType,
-                    types.BuiltinFunctionType,
-                    classmethod,
-                    staticmethod,
-                ),
-            )
-            and v is not None
-        }
-
-
-def validate_environment(api_key, headers: dict):
-    headers.update(
-        {
-            "Request-Source": "unspecified:litellm",
-            "accept": "application/json",
-            "content-type": "application/json",
-        }
-    )
-    if api_key:
-        headers["Authorization"] = f"Bearer {api_key}"
-    return headers
+from ...prompt_templates.factory import cohere_message_pt, cohere_messages_pt_v2
+from .transformation import CohereChatConfig, CohereError
 
 
 def translate_openai_tool_to_cohere(openai_tool):
@@ -321,7 +212,9 @@ def completion(  # noqa: PLR0915
     client=None,
     timeout=None,
 ):
-    headers = validate_environment(api_key, headers=headers)
+    headers = litellm.CohereChatConfig().validate_environment(
+        api_key=api_key, headers=headers, model=model, messages=messages
+    )
     completion_url = api_base
     model = model
     most_recent_message, chat_history = cohere_messages_pt_v2(
